+++ src/main/java/com/example/service/DistrictFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.District;
 
 public interface DistrictFetchService {
-	District fetchById(int id);
+	District fetchById(Long id);
 
 }
