+++ src/main/java/com/example/service/DistrictUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.District;
 
 public interface DistrictUpdateService {
-	District updateData(District district , int id);
+	District updateData(District district , Long id);
 
 }
